requirements.txt
python-debianbts==1.11
